PORT-22 Move navbar to app, remove menu in home and rename home to about and import about into App
